backend/python_service/requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-community
langchain-openai
chromadb
faiss-cpu
pymongo
openai
python-dotenv